+++ lightly_studio_view/src/lib/components/Histogram/index.ts
@@ -1,2 +1,2 @@
 export { default as Histogram } from './Histogram.svelte';
-export type { HistogramData, HistogramRange } from './types';
+export type { HistogramData, HistogramRange, HistogramSeries } from './types';